wb: add wbtest app
wbtest supports the v4l2 mem-2-mem api to test the OMAP WB M2M feature.
It loosely based on the vpetest app but built with libkms++ it also
support displaying the result on screen when required.

It can be invoked without any parameters in which case it will
create a test pattern frame buffer and convert it, but without
specifying "--display", the test will silently execute and terminate.

Signed-off-by: Benoit Parrot <bparrot@ti.com>